Método de listar terminado

+++ src/main/java/com/alura/literalura_challenge/model/DatosLibro.java
@@ -10,5 +10,5 @@ public record DatosLibro(
         @JsonAlias("title") String titulo,
         @JsonAlias("authors") List<DatosAutor> autor,
         @JsonAlias("languages") List<String> idiomas,
-        @JsonAlias("download_count") Double numeroDeDescargas
+        @JsonAlias("download_count") Integer numeroDeDescargas
 ) {}
